simplify the commit

+++ lib/fog/identity.rb
@@ -6,10 +6,18 @@ module Fog
     end
 
     def self.new(attributes)
-      attrs = attributes.dup # Prevent delete from having side effects
-      provider = attrs.delete(:provider).to_s.downcase.to_sym
-      raise ArgumentError.new("#{provider} has no identity service") unless self.providers.include?(provider)
-      Fog::Identity.const_get(Fog.providers[provider]).new(attrs)
+      attributes = attributes.dup # Prevent delete from having side effects
+      case provider = attributes.delete(:provider).to_s.downcase.to_sym
+      when :rackspace
+        require 'fog/rackspace/identity'
+        Fog::Rackspace::Identity.new(attributes)
+      else
+        if self.providers.include?(provider)
+          # require "fog/#{provider}/identity"
+          return Fog::Identity.const_get(Fog.providers[provider]).new(attributes)
+        end
+        raise ArgumentError.new("#{provider} has no identity service")
+      end
     end
 
     def self.providers
